+++ tests/integration/study_data_blueprint/test_st_storage.py
@@ -1,5 +1,6 @@
 import json
 import re
+import typing as t
 from unittest.mock import ANY
 
 import numpy as np
@@ -123,14 +124,15 @@ class TestSTStorage:
         # =============================
 
         # updating the matrix of a short-term storage
-        array = np.random.rand(8760, 1) * 1000
+        array = np.random.randint(0, 1000, size=(8760, 1))
+        array_list = array.tolist()
         res = client.put(
             f"/v1/studies/{study_id}/areas/{area_id}/storages/{siemens_battery_id}/series/inflows",
             headers={"Authorization": f"Bearer {user_access_token}"},
             json={
                 "index": list(range(array.shape[0])),
                 "columns": list(range(array.shape[1])),
-                "data": array.tolist(),
+                "data": array_list,
             },
         )
         assert res.status_code == 200, res.json()
@@ -232,6 +234,32 @@ class TestSTStorage:
         assert res.json() == siemens_config
 
         # =============================
+        #  SHORT-TERM STORAGE DUPLICATION
+        # =============================
+
+        new_name = "Duplicate of Siemens"
+        res = client.post(
+            f"/v1/studies/{study_id}/areas/{area_id}/storages/{siemens_battery_id}",
+            headers={"Authorization": f"Bearer {user_access_token}"},
+            params={"newName": new_name},
+        )
+        assert res.status_code in {200, 201}, res.json()
+        # asserts the config is the same
+        duplicated_config = dict(siemens_config)
+        duplicated_config["name"] = new_name  # type: ignore
+        duplicated_id = transform_name_to_id(new_name)
+        duplicated_config["id"] = duplicated_id  # type: ignore
+        assert res.json() == duplicated_config
+
+        # asserts the matrix has also been duplicated
+        res = client.get(
+            f"/v1/studies/{study_id}/areas/{area_id}/storages/{duplicated_id}/series/inflows",
+            headers={"Authorization": f"Bearer {user_access_token}"},
+        )
+        assert res.status_code == 200
+        assert res.json()["data"] == array_list
+
+        # =============================
         #  SHORT-TERM STORAGE DELETION
         # =============================
 
@@ -303,25 +331,25 @@ class TestSTStorage:
         assert res.status_code == 200, res.json()
         siemens_config = {**DEFAULT_PROPERTIES, **siemens_properties, "id": siemens_battery_id}
         grand_maison_config = {**DEFAULT_PROPERTIES, **grand_maison_properties, "id": grand_maison_id}
-        assert res.json() == [siemens_config, grand_maison_config]
+        assert res.json() == [duplicated_config, siemens_config, grand_maison_config]
 
-        # We can delete the two short-term storages at once.
+        # We can delete the three short-term storages at once.
         res = client.request(
             "DELETE",
             f"/v1/studies/{study_id}/areas/{area_id}/storages",
             headers={"Authorization": f"Bearer {user_access_token}"},
-            json=[siemens_battery_id, grand_maison_id],
+            json=[grand_maison_id, duplicated_config["id"]],
         )
         assert res.status_code == 204, res.json()
         assert res.text in {"", "null"}  # Old FastAPI versions return 'null'.
 
-        # The list of short-term storages should be empty.
+        # Only one st-storage should remain.
         res = client.get(
             f"/v1/studies/{study_id}/areas/{area_id}/storages",
             headers={"Authorization": f"Bearer {user_access_token}"},
         )
         assert res.status_code == 200, res.json()
-        assert res.json() == []
+        assert len(res.json()) == 1
 
         # ===========================
         #  SHORT-TERM STORAGE ERRORS
@@ -449,6 +477,30 @@ class TestSTStorage:
         obj = res.json()
         description = obj["description"]
         assert bad_study_id in description
+
+        # Cannot duplicate a fake st-storage
+        unknown_id = "unknown"
+        res = client.post(
+            f"/v1/studies/{study_id}/areas/{area_id}/storages/{unknown_id}",
+            headers={"Authorization": f"Bearer {user_access_token}"},
+            params={"newName": "duplicata"},
+        )
+        assert res.status_code == 404, res.json()
+        obj = res.json()
+        assert obj["description"] == f"Fields of storage '{unknown_id}' not found"
+        assert obj["exception"] == "STStorageFieldsNotFoundError"
+
+        # Cannot duplicate with an existing id
+        res = client.post(
+            f"/v1/studies/{study_id}/areas/{area_id}/storages/{siemens_battery_id}",
+            headers={"Authorization": f"Bearer {user_access_token}"},
+            params={"newName": siemens_battery.upper()},  # different case, but same ID
+        )
+        assert res.status_code == 409, res.json()
+        obj = res.json()
+        description = obj["description"]
+        assert siemens_battery.lower() in description
+        assert obj["exception"] == "ClusterAlreadyExists"
 
     def test__default_values(
         self,
@@ -632,3 +684,146 @@ class TestSTStorage:
             "initiallevel": 0.0,
         }
         assert actual == expected
+
+    @pytest.fixture(name="base_study_id")
+    def base_study_id_fixture(self, request: t.Any, client: TestClient, user_access_token: str) -> str:
+        """Prepare a managed study for the variant study tests."""
+        params = request.param
+        res = client.post(
+            "/v1/studies",
+            headers={"Authorization": f"Bearer {user_access_token}"},
+            params=params,
+        )
+        assert res.status_code in {200, 201}, res.json()
+        study_id: str = res.json()
+        return study_id
+
+    @pytest.fixture(name="variant_id")
+    def variant_id_fixture(self, request: t.Any, client: TestClient, user_access_token: str, base_study_id: str) -> str:
+        """Prepare a variant study for the variant study tests."""
+        name = request.param
+        res = client.post(
+            f"/v1/studies/{base_study_id}/variants",
+            headers={"Authorization": f"Bearer {user_access_token}"},
+            params={"name": name},
+        )
+        assert res.status_code in {200, 201}, res.json()
+        study_id: str = res.json()
+        return study_id
+
+    # noinspection PyTestParametrized
+    @pytest.mark.parametrize("base_study_id", [{"name": "Base Study", "version": 860}], indirect=True)
+    @pytest.mark.parametrize("variant_id", ["Variant Study"], indirect=True)
+    def test_variant_lifecycle(self, client: TestClient, user_access_token: str, variant_id: str) -> None:
+        """
+        In this test, we want to check that short-term storages can be managed
+        in the context of a "variant" study.
+        """
+        # Create an area
+        area_name = "France"
+        res = client.post(
+            f"/v1/studies/{variant_id}/areas",
+            headers={"Authorization": f"Bearer {user_access_token}"},
+            json={"name": area_name, "type": "AREA"},
+        )
+        assert res.status_code in {200, 201}, res.json()
+        area_cfg = res.json()
+        area_id = area_cfg["id"]
+
+        # Create a short-term storage
+        cluster_name = "Tesla1"
+        res = client.post(
+            f"/v1/studies/{variant_id}/areas/{area_id}/storages",
+            headers={"Authorization": f"Bearer {user_access_token}"},
+            json={
+                "name": cluster_name,
+                "group": "Battery",
+                "injectionNominalCapacity": 4500,
+                "withdrawalNominalCapacity": 4230,
+                "reservoirCapacity": 5700,
+            },
+        )
+        assert res.status_code in {200, 201}, res.json()
+        cluster_id: str = res.json()["id"]
+
+        # Update the short-term storage
+        res = client.patch(
+            f"/v1/studies/{variant_id}/areas/{area_id}/storages/{cluster_id}",
+            headers={"Authorization": f"Bearer {user_access_token}"},
+            json={"reservoirCapacity": 5600},
+        )
+        assert res.status_code == 200, res.json()
+        cluster_cfg = res.json()
+        assert cluster_cfg["reservoirCapacity"] == 5600
+
+        # Update the series matrix
+        matrix = np.random.randint(0, 2, size=(8760, 1)).tolist()
+        matrix_path = f"input/st-storage/series/{area_id}/{cluster_id.lower()}/pmax_injection"
+        args = {"target": matrix_path, "matrix": matrix}
+        res = client.post(
+            f"/v1/studies/{variant_id}/commands",
+            json=[{"action": "replace_matrix", "args": args}],
+            headers={"Authorization": f"Bearer {user_access_token}"},
+        )
+        assert res.status_code in {200, 201}, res.json()
+
+        # Duplicate the short-term storage
+        new_name = "Tesla2"
+        res = client.post(
+            f"/v1/studies/{variant_id}/areas/{area_id}/storages/{cluster_id}",
+            headers={"Authorization": f"Bearer {user_access_token}"},
+            params={"newName": new_name},
+        )
+        assert res.status_code in {200, 201}, res.json()
+        cluster_cfg = res.json()
+        assert cluster_cfg["name"] == new_name
+        new_id = cluster_cfg["id"]
+
+        # Check that the duplicate has the right properties
+        res = client.get(
+            f"/v1/studies/{variant_id}/areas/{area_id}/storages/{new_id}",
+            headers={"Authorization": f"Bearer {user_access_token}"},
+        )
+        assert res.status_code == 200, res.json()
+        cluster_cfg = res.json()
+        assert cluster_cfg["group"] == "Battery"
+        assert cluster_cfg["injectionNominalCapacity"] == 4500
+        assert cluster_cfg["withdrawalNominalCapacity"] == 4230
+        assert cluster_cfg["reservoirCapacity"] == 5600
+
+        # Check that the duplicate has the right matrix
+        new_cluster_matrix_path = f"input/st-storage/series/{area_id}/{new_id.lower()}/pmax_injection"
+        res = client.get(
+            f"/v1/studies/{variant_id}/raw",
+            params={"path": new_cluster_matrix_path},
+            headers={"Authorization": f"Bearer {user_access_token}"},
+        )
+        assert res.status_code == 200
+        assert res.json()["data"] == matrix
+
+        # Delete the short-term storage
+        res = client.delete(
+            f"/v1/studies/{variant_id}/areas/{area_id}/storages",
+            headers={"Authorization": f"Bearer {user_access_token}"},
+            json=[cluster_id],
+        )
+        assert res.status_code == 204, res.json()
+
+        # Check the list of variant commands
+        res = client.get(
+            f"/v1/studies/{variant_id}/commands",
+            headers={"Authorization": f"Bearer {user_access_token}"},
+        )
+        assert res.status_code == 200, res.json()
+        commands = res.json()
+        assert len(commands) == 7
+        actions = [command["action"] for command in commands]
+        assert actions == [
+            "create_area",
+            "create_st_storage",
+            "update_config",
+            "replace_matrix",
+            "create_st_storage",
+            "replace_matrix",
+            "remove_st_storage",
+        ]